requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
flask-restx==1.3.0
frozenlist==1.4.1
graphql-core==3.2.3
h11==0.14.0
httptools==0.6.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
libcst==1.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
openpyxl==3.1.4
passlib==1.7.4
phonenumbers==8.13.39
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
rich==13.7.1
Routes==2.5.1
rpds-py==0.18.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.235.0
twilio==9.2.0
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
